add compat for ctv codec
