+++ src/main/scala/ai/metarank/fstore/codec/impl/ItemIdCodec.scala
@@ -1,0 +1,10 @@
+package ai.metarank.fstore.codec.impl
+
+import ai.metarank.model.Identifier.ItemId
+
+import java.io.{DataInput, DataOutput}
+
+object ItemIdCodec extends BinaryCodec[ItemId] {
+  override def read(in: DataInput): ItemId = ItemId(in.readUTF())
+  override def write(value: ItemId, out: DataOutput): Unit = out.writeUTF(value.value)
+}
