+++ src/test/resources/codec/ctv-v3.bin
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:1cd75fda1f292507c2f805e1d8d90782d808c68e89729244bb7f2f6d983020aa
+size 210
